--- Assets/02.Script/Enmey/DetectPlayerCollider.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: f8b18453dd5b7d6428ed5744ed54dcb5
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
