+++ Assets/02.Script/Enmey/EnemyArea.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 3f7984aa459eda64b8d732d69c1c7386
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
