+++ Assets/02.Script/Enmey/EnemyArea/MonsterAreaCollider.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: da25be588edbbdc46b10c60e84ec171d
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
