+++ Assets/02.Script/Enmey/EnemyArea/RespawnController.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 078ff5103e8fc8542a5946645238de12
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
