+++ Assets/02.Script/Enmey/EnemyArea/RespawnPoint.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 335c3ff3201961144893fbed6c32026d
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
