--- Assets/02.Script/Enmey/EnemyAttackArea.cs
@@ -1,31 +1,0 @@
-using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-
-public class EnemyAttackArea : MonoBehaviour
-{
-    [SerializeField]
-    private Enemy enemy;
-
-    private void OnTriggerEnter(Collider other)
-    {
-        if (other.CompareTag("Player"))
-        {
-            enemy.EnemyAttack();
-        }
-    }
-    private void OnTriggerStay(Collider other)
-    {
-        if (other.CompareTag("Player"))
-        {
-            enemy.EnemyAttack();
-        }
-    }
-    private void OnTriggerExit(Collider other)
-    {
-        if (other.CompareTag("Player"))
-        {
-            enemy.EnemyMoveToPlayer();
-        }
-    }
-}
