--- Assets/02.Script/Enmey/EnemyAttackArea.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 09e52a3056f6daf45a00fdd72b8a308e
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
