--- Assets/02.Script/Enmey/EnemyCanvasController.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: fd8a00514aecb954e97b3d382b4d7ecc
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
