+++ Assets/02.Script/Enmey/EnemyPublic.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: bc1c0ca852032dd4eb213c8d57513d4e
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
