+++ Assets/02.Script/Enmey/EnemyPublic/EnemyAttackArea.cs
@@ -1,0 +1,31 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+public class EnemyAttackArea : MonoBehaviour
+{
+    [SerializeField]
+    private Enemy enemy;
+
+    private void OnTriggerEnter(Collider other)
+    {
+        if (other.CompareTag("Player"))
+        {
+            enemy.EnemyAttack();
+        }
+    }
+    private void OnTriggerStay(Collider other)
+    {
+        if (other.CompareTag("Player"))
+        {
+            enemy.EnemyAttack();
+        }
+    }
+    private void OnTriggerExit(Collider other)
+    {
+        if (other.CompareTag("Player"))
+        {
+            enemy.EnemyMoveToPlayer();
+        }
+    }
+}
