+++ Assets/02.Script/Enmey/EnemyPublic/EnemyCollider.cs
@@ -1,0 +1,12 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+public class EnemyCollider : MonoBehaviour
+{
+    [SerializeField]
+    private Enemy currentEnemy;
+
+    public void CheckMosterArea(bool isArea) => currentEnemy.CheckMonsterArea(isArea);
+    public void DetectPlayer(bool detect) => currentEnemy.DetectPlayer(detect);
+}
