+++ Assets/02.Script/Enmey/EnemyPublic/EnemyCollider.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 603563e8a9052a64693cc6658fa76f41
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
