+++ Assets/02.Script/Enmey/EnemyPublic/EnemyHPBar.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 6bff37bf745eb974a83bd175f73b10a4
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
