+++ Assets/02.Script/Enmey/EnemyPublic/EnemySpawnPosition.cs
@@ -1,0 +1,17 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+public class EnemySpawnPosition : MonoBehaviour
+{
+    [SerializeField]
+    public Enemy enemy;
+
+    private void OnTriggerStay(Collider other)
+    {
+        if (other.CompareTag("Enemy"))
+        {
+            enemy.StaySpawnPosition();
+        }        
+    }
+}
