--- Assets/02.Script/Enmey/EnemySpawnPosition.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 2264e4a8ed80b264abde8557dc868c06
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
