+++ Assets/04.Objects/Enemy/EvilMage.prefab
@@ -15,7 +15,7 @@ GameObject:
   m_Icon: {fileID: 0}
   m_NavMeshLayer: 0
   m_StaticEditorFlags: 0
-  m_IsActive: 0
+  m_IsActive: 1
 --- !u!4 &6361552430185234492
 Transform:
   m_ObjectHideFlags: 0
@@ -162,6 +162,10 @@ PrefabInstance:
       value: 30
       objectReference: {fileID: 0}
     - target: {fileID: 3983735538642761398, guid: 858dd493f7e497a4e992fb2d1e11ad4e, type: 3}
+      propertyPath: animator
+      value: 
+      objectReference: {fileID: 5543008952003900543}
+    - target: {fileID: 3983735538642761398, guid: 858dd493f7e497a4e992fb2d1e11ad4e, type: 3}
       propertyPath: spawnPosition
       value: 
       objectReference: {fileID: 4523783828419757048}
@@ -172,6 +176,11 @@ PrefabInstance:
     m_RemovedComponents: []
     m_AddedGameObjects: []
   m_SourcePrefab: {fileID: 100100000, guid: 858dd493f7e497a4e992fb2d1e11ad4e, type: 3}
+--- !u!95 &5543008952003900543 stripped
+Animator:
+  m_CorrespondingSourceObject: {fileID: 3284327623693519156, guid: 858dd493f7e497a4e992fb2d1e11ad4e, type: 3}
+  m_PrefabInstance: {fileID: 7023644907218795851}
+  m_PrefabAsset: {fileID: 0}
 --- !u!1 &5543008952013408161 stripped
 GameObject:
   m_CorrespondingSourceObject: {fileID: 3284327623702887146, guid: 858dd493f7e497a4e992fb2d1e11ad4e, type: 3}
