+++ Assets/04.Objects/Enemy/Orc.prefab
@@ -25,7 +25,7 @@ Transform:
   m_PrefabAsset: {fileID: 0}
   m_GameObject: {fileID: 802534884909208570}
   m_LocalRotation: {x: 0, y: 0, z: 0, w: 1}
-  m_LocalPosition: {x: 0, y: 10, z: 0}
+  m_LocalPosition: {x: 0, y: 0, z: 10}
   m_LocalScale: {x: 10, y: 10, z: 1}
   m_ConstrainProportionsScale: 0
   m_Children: []
@@ -126,6 +126,10 @@ PrefabInstance:
     m_TransformParent: {fileID: 5475433763304560733}
     m_Modifications:
     - target: {fileID: 1576408255903471720, guid: 9ee7f74360f530b4ab0ab58080468ed1, type: 3}
+      propertyPath: animator
+      value: 
+      objectReference: {fileID: 9125284091870137128}
+    - target: {fileID: 1576408255903471720, guid: 9ee7f74360f530b4ab0ab58080468ed1, type: 3}
       propertyPath: orcStatus
       value: 
       objectReference: {fileID: 4751397622702492100}
@@ -214,6 +218,11 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: aa5f066058a19fd46a1088cb39258071, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
+--- !u!95 &9125284091870137128 stripped
+Animator:
+  m_CorrespondingSourceObject: {fileID: 7513959790142719054, guid: 9ee7f74360f530b4ab0ab58080468ed1, type: 3}
+  m_PrefabInstance: {fileID: 1649855779009187686}
+  m_PrefabAsset: {fileID: 0}
 --- !u!1 &9125284091877406966 stripped
 GameObject:
   m_CorrespondingSourceObject: {fileID: 7513959790152251280, guid: 9ee7f74360f530b4ab0ab58080468ed1, type: 3}
