+++ Assets/04.Objects/Enemy/Slime.prefab
@@ -9,13 +9,14 @@ GameObject:
   serializedVersion: 6
   m_Component:
   - component: {fileID: 3181034117998330993}
+  - component: {fileID: 2552185466461477439}
   m_Layer: 0
   m_Name: Slime
-  m_TagString: Untagged
+  m_TagString: RespawnEnemy
   m_Icon: {fileID: 0}
   m_NavMeshLayer: 0
   m_StaticEditorFlags: 0
-  m_IsActive: 0
+  m_IsActive: 1
 --- !u!4 &3181034117998330993
 Transform:
   m_ObjectHideFlags: 0
@@ -28,11 +29,108 @@ Transform:
   m_LocalScale: {x: 1, y: 1, z: 1}
   m_ConstrainProportionsScale: 0
   m_Children:
-  - {fileID: 4664507495873121769}
   - {fileID: 1837110710956500584}
+  - {fileID: 4664507495873121769}
   m_Father: {fileID: 0}
-  m_RootOrder: 15
+  m_RootOrder: 38
   m_LocalEulerAnglesHint: {x: 0, y: 180, z: 0}
+--- !u!65 &2552185466461477439
+BoxCollider:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 3169693915318950036}
+  m_Material: {fileID: 0}
+  m_IsTrigger: 1
+  m_Enabled: 1
+  serializedVersion: 2
+  m_Size: {x: 1, y: 1, z: 1}
+  m_Center: {x: 0, y: 0, z: 0}
+--- !u!1 &6760323635127458509
+GameObject:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  serializedVersion: 6
+  m_Component:
+  - component: {fileID: 2980416547397573687}
+  - component: {fileID: 4622801241827484295}
+  m_Layer: 10
+  m_Name: MiniMapPoint
+  m_TagString: Untagged
+  m_Icon: {fileID: 0}
+  m_NavMeshLayer: 0
+  m_StaticEditorFlags: 0
+  m_IsActive: 1
+--- !u!4 &2980416547397573687
+Transform:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 6760323635127458509}
+  m_LocalRotation: {x: -0, y: -0, z: -1, w: 0}
+  m_LocalPosition: {x: 0, y: 0, z: 10}
+  m_LocalScale: {x: 10, y: 10, z: 1}
+  m_ConstrainProportionsScale: 0
+  m_Children: []
+  m_Father: {fileID: 1837371722408735530}
+  m_RootOrder: -1
+  m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
+--- !u!212 &4622801241827484295
+SpriteRenderer:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 6760323635127458509}
+  m_Enabled: 1
+  m_CastShadows: 0
+  m_ReceiveShadows: 0
+  m_DynamicOccludee: 1
+  m_StaticShadowCaster: 0
+  m_MotionVectors: 1
+  m_LightProbeUsage: 1
+  m_ReflectionProbeUsage: 1
+  m_RayTracingMode: 0
+  m_RayTraceProcedural: 0
+  m_RenderingLayerMask: 1
+  m_RendererPriority: 0
+  m_Materials:
+  - {fileID: 10754, guid: 0000000000000000f000000000000000, type: 0}
+  m_StaticBatchInfo:
+    firstSubMesh: 0
+    subMeshCount: 0
+  m_StaticBatchRoot: {fileID: 0}
+  m_ProbeAnchor: {fileID: 0}
+  m_LightProbeVolumeOverride: {fileID: 0}
+  m_ScaleInLightmap: 1
+  m_ReceiveGI: 1
+  m_PreserveUVs: 0
+  m_IgnoreNormalsForChartDetection: 0
+  m_ImportantGI: 0
+  m_StitchLightmapSeams: 1
+  m_SelectedEditorRenderState: 0
+  m_MinimumChartSize: 4
+  m_AutoUVMaxDistance: 0.5
+  m_AutoUVMaxAngle: 89
+  m_LightmapParameters: {fileID: 0}
+  m_SortingLayerID: 0
+  m_SortingLayer: 0
+  m_SortingOrder: 0
+  m_Sprite: {fileID: 21300000, guid: 46d311c34f66a4d7ab285ebac88de2ca, type: 3}
+  m_Color: {r: 1, g: 0, b: 0.1218915, a: 1}
+  m_FlipX: 0
+  m_FlipY: 0
+  m_DrawMode: 0
+  m_Size: {x: 0.9, y: 0.9}
+  m_AdaptiveModeThreshold: 0.5
+  m_SpriteTileMode: 0
+  m_WasSpriteAssigned: 1
+  m_MaskInteraction: 0
+  m_SpriteSortPoint: 0
 --- !u!1001 &2526647696069612836
 PrefabInstance:
   m_ObjectHideFlags: 0
@@ -41,6 +139,18 @@ PrefabInstance:
     serializedVersion: 2
     m_TransformParent: {fileID: 3181034117998330993}
     m_Modifications:
+    - target: {fileID: 1099833740646510211, guid: 17288e7ab25d2394fa2f1f5a8b4fb5de, type: 3}
+      propertyPath: enemyStatus
+      value: 
+      objectReference: {fileID: 2937067375066434912}
+    - target: {fileID: 1202430507243679485, guid: 17288e7ab25d2394fa2f1f5a8b4fb5de, type: 3}
+      propertyPath: animator
+      value: 
+      objectReference: {fileID: 1739508690819764016}
+    - target: {fileID: 1202430507243679485, guid: 17288e7ab25d2394fa2f1f5a8b4fb5de, type: 3}
+      propertyPath: slimeStatus
+      value: 
+      objectReference: {fileID: 2937067375066434912}
     - target: {fileID: 1202430507243679485, guid: 17288e7ab25d2394fa2f1f5a8b4fb5de, type: 3}
       propertyPath: spawnPosition
       value: 
@@ -94,8 +204,16 @@ PrefabInstance:
       value: 0
       objectReference: {fileID: 0}
     m_RemovedComponents: []
-    m_AddedGameObjects: []
+    m_AddedGameObjects:
+    - targetCorrespondingSourceObject: {fileID: 4210816627195569678, guid: 17288e7ab25d2394fa2f1f5a8b4fb5de, type: 3}
+      insertIndex: -1
+      addedObject: {fileID: 2980416547397573687}
   m_SourcePrefab: {fileID: 100100000, guid: 17288e7ab25d2394fa2f1f5a8b4fb5de, type: 3}
+--- !u!95 &1739508690819764016 stripped
+Animator:
+  m_CorrespondingSourceObject: {fileID: 4265908584456231444, guid: 17288e7ab25d2394fa2f1f5a8b4fb5de, type: 3}
+  m_PrefabInstance: {fileID: 2526647696069612836}
+  m_PrefabAsset: {fileID: 0}
 --- !u!1 &1834662074768319276 stripped
 GameObject:
   m_CorrespondingSourceObject: {fileID: 4208178301790565896, guid: 17288e7ab25d2394fa2f1f5a8b4fb5de, type: 3}
@@ -202,6 +320,22 @@ Transform:
   m_CorrespondingSourceObject: {fileID: 4210530043239691084, guid: 17288e7ab25d2394fa2f1f5a8b4fb5de, type: 3}
   m_PrefabInstance: {fileID: 2526647696069612836}
   m_PrefabAsset: {fileID: 0}
+--- !u!4 &1837371722408735530 stripped
+Transform:
+  m_CorrespondingSourceObject: {fileID: 4210816627195569678, guid: 17288e7ab25d2394fa2f1f5a8b4fb5de, type: 3}
+  m_PrefabInstance: {fileID: 2526647696069612836}
+  m_PrefabAsset: {fileID: 0}
+--- !u!114 &2937067375066434912 stripped
+MonoBehaviour:
+  m_CorrespondingSourceObject: {fileID: 852019833425926212, guid: 17288e7ab25d2394fa2f1f5a8b4fb5de, type: 3}
+  m_PrefabInstance: {fileID: 2526647696069612836}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 1834662074768319276}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: af7c44888390db941a1d4cb29e33cfb2, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
 --- !u!114 &3728858023416036313 stripped
 MonoBehaviour:
   m_CorrespondingSourceObject: {fileID: 1202430507243679485, guid: 17288e7ab25d2394fa2f1f5a8b4fb5de, type: 3}
