--- Assets/04.Objects/EnemyArea.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 9fdb527c6856e9645926e71c50f6e130
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
