--- Assets/04.Objects/EnemyArea/EnemyArea.prefab.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 9835644a99e7d2241b2e594a793c0e49
-PrefabImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
