+++ Assets/04.Objects/RespaenArea/EnemyArea.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 9835644a99e7d2241b2e594a793c0e49
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
