+++ Assets/04.Objects/RespaenArea/RespawnPoint.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 64e9c836bed0d17448aa30dae3a9bf60
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
