+++ Assets/Rhinox Studio/RPG - Stylized Fantasy Environment Built-In/terrain/Terrain.asset
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:632d177b2a1e0fc9028a7c611a658411c1cb1c3ba83158a2dd4c3311208ac445
-size 145694348
+oid sha256:7917d5cbf822c1910d73fd93441270f84d2278c4008f5f591f4d55dade39cdf3
+size 145693804
